fix: 로컬환경용 멤버 설정 제거
Former-commit-id: b2789983074bea299d48909d58150df55458db9c

+++ backend/src/main/java/com/tfeo/backend/domain/member/controller/MemberController.java
@@ -130,13 +130,12 @@ public class MemberController {
 	//회원 상세정보 조회
 	@GetMapping("")
 	public ResponseEntity<?> memberDetails(HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L;
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		MemberResponseDto memberResponseDto = memberService.findMember(memberNo);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
@@ -150,13 +149,12 @@ public class MemberController {
 	@PostMapping("/survey")
 	public ResponseEntity<?> memberSurvey(@RequestBody SurveyRequestDto memberSurveyRequestDto,
 		HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L;
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		memberService.submitSurvey(memberSurveyRequestDto, memberNo);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
@@ -169,12 +167,11 @@ public class MemberController {
 	@PutMapping("")
 	public ResponseEntity<?> memberModify(@RequestBody MemberRequestDto memberRequestDto, HttpServletRequest request) {
 		Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L;
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		return ResponseEntity.ok(SuccessResponse.builder()
 			.status(HttpStatus.OK)
 			.result(memberService.modifyMember(memberNo, memberRequestDto))
@@ -200,13 +197,12 @@ public class MemberController {
 	// 회원이 신청한 집(계약) 조회
 	@GetMapping("/home")
 	public ResponseEntity<?> AppliedHomeList(HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L; //Todo: contract 개발 완료 이후 제거
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		AppliedHomeResponseDto appliedHomeResponseDto = memberService.findAppliedHome(memberNo);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
@@ -219,13 +215,12 @@ public class MemberController {
 	@PostMapping("/home/{homeNo}")
 	public ResponseEntity<?> homeApplicationAdd(@PathVariable Long homeNo,
 		@RequestBody MemberHomeApplicationRequestDto memberHomeApplicationRequestDto, HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L; // Todo: 프런트 개발 이후 제거
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		memberService.addHomeApplication(homeNo, memberNo, memberHomeApplicationRequestDto);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
@@ -237,13 +232,12 @@ public class MemberController {
 	//집 신청 취소
 	@DeleteMapping("/home/{homeNo}")
 	public ResponseEntity<?> memberRequestedHomeRemove(@PathVariable Long homeNo, HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L; // Todo: 프런트 개발 이후 제거
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		memberService.deleteApplication(homeNo, memberNo);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
